refactor(ui): extract coordinator-admin/components/devices-panel.ts
Move renderDevicesPanel into its own module. The panel now takes a deps
object with { summary, runDevice } so the barrel wires in the factory
action. The last use of TextInput in coordinator-admin.tsx also moves out.
